Add bndrun file with proper configuration
We need to specify several things for OSGi integration testing and we
need to use a bndrun file since the resolve task will compute and
write the `-runbundles` instruction to be used by the testOSGi task. So
we need a bndrun file to capture thsi inter-task information. Generally,
you will want to commit any change from resolve to the bndrun file back
to git.

We update the build.gradle file to add the necessary dependencies and
put them in the proper configurations. We also need to make sure the
test bundle has a different Bundle-SymbolicName from the main bundle.
Bnd 5.1.2 does not do this, but I will make a fix to Bnd 5.2.0.

We add the resolve task and make sure the testOSGi task runs after the
resolve task. We don't always need to run the resolve task just to
rerun the testOSGi task, but we have the check task depend upon both.

Now running the build task will build the project, run plain JUnit tests
and also run the OSGi integration tests.